requirements-dev.txt
asgiref==3.7.2
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
defusedxml==0.7.1
distro==1.7.0
Django==5.0.2
django-allauth==0.61.1
django-environ==0.11.2
django-widgets-improved==1.5.0
fonttools==4.47.2
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
matplotlib==3.8.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
packaging==23.2
pillow==10.2.0
pycparser==2.22
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
wadllib==1.3.6
zipp==1.0.0
